LPS-75989 Move custom.attributes to internal.
