+++ modules/apps/foundation/users-admin/users-admin-web/src/main/java/com/liferay/users/admin/web/internal/custom/attributes/LayoutCustomAttributesDisplay.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.users.admin.web.custom.attributes;
+package com.liferay.users.admin.web.internal.custom.attributes;
 
 import com.liferay.expando.kernel.model.BaseCustomAttributesDisplay;
 import com.liferay.expando.kernel.model.CustomAttributesDisplay;
-import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.model.Layout;
 import com.liferay.users.admin.constants.UsersAdminPortletKeys;
 
 import org.osgi.service.component.annotations.Component;
@@ -29,11 +29,16 @@ import org.osgi.service.component.annotations.Component;
 	property = {"javax.portlet.name=" + UsersAdminPortletKeys.USERS_ADMIN},
 	service = CustomAttributesDisplay.class
 )
-public class UserCustomAttributesDisplay extends BaseCustomAttributesDisplay {
+public class LayoutCustomAttributesDisplay extends BaseCustomAttributesDisplay {
 
 	@Override
 	public String getClassName() {
-		return User.class.getName();
+		return Layout.class.getName();
+	}
+
+	@Override
+	public String getIconCssClass() {
+		return "edit-layout";
 	}
 
 }
