+++ modules/apps/foundation/users-admin/users-admin-web/src/main/java/com/liferay/users/admin/web/internal/custom/attributes/OrganizationCustomAttributesDisplay.java
@@ -12,11 +12,11 @@
  * details.
  */
 
-package com.liferay.users.admin.web.custom.attributes;
+package com.liferay.users.admin.web.internal.custom.attributes;
 
 import com.liferay.expando.kernel.model.BaseCustomAttributesDisplay;
 import com.liferay.expando.kernel.model.CustomAttributesDisplay;
-import com.liferay.portal.kernel.model.Layout;
+import com.liferay.portal.kernel.model.Organization;
 import com.liferay.users.admin.constants.UsersAdminPortletKeys;
 
 import org.osgi.service.component.annotations.Component;
@@ -29,16 +29,17 @@ import org.osgi.service.component.annotations.Component;
 	property = {"javax.portlet.name=" + UsersAdminPortletKeys.USERS_ADMIN},
 	service = CustomAttributesDisplay.class
 )
-public class LayoutCustomAttributesDisplay extends BaseCustomAttributesDisplay {
+public class OrganizationCustomAttributesDisplay
+	extends BaseCustomAttributesDisplay {
 
 	@Override
 	public String getClassName() {
-		return Layout.class.getName();
+		return Organization.class.getName();
 	}
 
 	@Override
 	public String getIconCssClass() {
-		return "edit-layout";
+		return "sites";
 	}
 
 }
