--- modules/apps/foundation/users-admin/users-admin-web/src/main/resources/com/liferay/users/admin/web/custom/attributes/packageinfo
@@ -1,1 +1,0 @@
-version 1.0.0
